requirements.txt
flask
flask-sqlalchemy
flask-login
werkzeug
tensorflow
numpy
requests
psycopg2-binary
python-dotenv
scikit-learn
